Modularized actuator code. 	modified:   actuator_pwm.py
	modified:   config.json
