+++ python/config.json
@@ -6,7 +6,7 @@
 	},
 
 	"distInfo" : {
-		"distActuatorToOrigin" : 3.5
+		"distActuatorToOrigin" : 2.8
 	}
 
 }
